web/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baboom"
version = "0.1.0"
description = ""
requires-python = ">= 3.14"

dependencies = [
  "gunicorn",
  "whitenoise",
  "psycopg",
  "requests",
  "Pillow",
  "django",
  "django-filter",
  "django-simple-history",
  "django-treebeard",
  "django-nested-admin",
  "django-environ",
  "django-htmx",
  "django-tailwind-cli",
  "django-widget-tweaks",
  "heroicons[django]",
]

[project.urls]
Repository = "https://github.com/rflemerson/baboom"

[project.optional-dependencies]
dev = [
  "mypy",
  "django-stubs",
  "django-stubs-ext",
  "ruff",
  "pre-commit",
]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".venv",
    ".git",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "DJ",  # django specific
    "B",   # flake8-bugbear
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"baboom/settings/__init__.py" = ["F403", "F405"]
"core/models.py" = ["DJ001"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["^baboom/"]

[tool.django-stubs]
django_settings_module = "baboom.settings.base"

[[tool.mypy.overrides]]
module = ["baboom.*"]
ignore_errors = true
